latent-loop-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latent-loop-mcp"
version = "0.1.0"
description = "MCP server for inspectable recurrent-style reasoning control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
latent-loop-mcp = "latent_loop_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/latent_loop_mcp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
